ref: Updates to align with other SDKs
