+++ src/Tracing/SamplingContext.php
@@ -17,6 +17,12 @@ final class SamplingContext
     private $parentSampled;
 
     /**
+     * @var array<string, mixed>|null Additional context, depending on where the SDK runs
+     */
+    private $additionalContext;
+
+
+    /**
      * Returns an instance populated with the data of the transaction context.
      */
     public static function getDefault(TransactionContext $transactionContext): self
@@ -47,5 +53,21 @@ final class SamplingContext
     public function setParentSampled(?bool $parentSampled): void
     {
         $this->parentSampled = $parentSampled;
+    }
+
+    /**
+     * @param array<string, mixed>|null $additionalContext
+     */
+    public function setAdditionalContext(?array $additionalContext): void
+    {
+        $this->additionalContext = $additionalContext;
+    }
+
+    /**
+     * @return array<string, mixed>|null
+     */
+    public function getAdditionalContext(): ?array
+    {
+        return $this->additionalContext;
     }
 }
